Fixed user profile edit form showing as empty on redirect
